% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{download_button}
\alias{download_button}
\title{FUNCTION_TITLE}
\usage{
download_button(
  inputId = "btn-Convert-Html2Image",
  label = "Download",
  contentId = "html-content-holder",
  filename = "canvas.png"
)
}
\arguments{
\item{inputId}{PARAM_DESCRIPTION, Default: 'btn-Convert-Html2Image'}

\item{label}{PARAM_DESCRIPTION, Default: 'Download'}

\item{contentId}{PARAM_DESCRIPTION, Default: 'html-content-holder'}

\item{filename}{PARAM_DESCRIPTION, Default: 'canvas.png'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{actionButton}}
}
