% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{preview_button}
\alias{preview_button}
\title{FUNCTION_TITLE}
\usage{
preview_button(
  inputId = "btn-Preview-Image",
  label = "Preview",
  contentId = "html-content-holder",
  previewId = "previewImage"
)
}
\arguments{
\item{inputId}{PARAM_DESCRIPTION, Default: 'btn-Preview-Image'}

\item{label}{PARAM_DESCRIPTION, Default: 'Preview'}

\item{contentId}{PARAM_DESCRIPTION, Default: 'html-content-holder'}

\item{previewId}{PARAM_DESCRIPTION, Default: 'previewImage'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{actionButton}}
}
