% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_snapper}
\alias{load_snapper}
\title{FUNCTION_TITLE}
\usage{
load_snapper()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{tag}}, \code{\link[shiny]{builder}}
}
