% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divs.R
\name{snapper_div}
\alias{snapper_div}
\title{FUNCTION_TITLE}
\usage{
snapper_div(id = "previewImage", ...)
}
\arguments{
\item{id}{PARAM_DESCRIPTION, Default: 'previewImage'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{builder}}
}
